split gatsby-node into separate files
